chore: move the useLockScroll file to the hook folder
